# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Recipe Catalogue API
  description: >
    Stores recipes and the reusable tags attached to them. A recipe has ordered
    steps (each carrying the tools that step needs), the catalogue entries it uses
    (each optionally with a quantity, and marked optional and/or replaceable with
    named substitutes), and tags. Recipes that share a name are variations, keyed
    on (name, creator, version). Cross-service references (ingredientId,
    replacement recipeId) are ids only, no foreign keys. Hand-maintained
    contract; keep it in step with the controllers under
    `com.example.recipecatalogue.Controllers`.

    Tag conventions: "recipe-exclusive" facets are ordinary tags, namespaced -
    `tool:beater`, `difficulty:easy`, `time:90min`, `category:snack`. Tags
    inherited from an entry's foods/ingredients are resolved by the BFF, not
    here.
  version: 0.0.2
servers:
  - url: http://localhost:8083
    description: Local

tags:
  - name: recipes
  - name: tags

paths:
  /api/recipes:
    get:
      tags: [recipes]
      summary: List recipes; every supplied filter is ANDed
      parameters:
        - name: name
          in: query
          description: Case-insensitive substring of the recipe name.
          schema: { type: string }
        - name: tag
          in: query
          description: Repeatable. Tag names to filter by (case-insensitive).
          schema: { type: array, items: { type: string } }
          example: ["cuisine:thai", "quick"]
        - name: match
          in: query
          description: '`all` (default) requires every tag; `any` requires at least one.'
          schema: { type: string, enum: [all, any], default: all }
        - name: ingredientId
          in: query
          description: Repeatable. Recipes that use any of these catalogue entries.
          schema: { type: array, items: { type: integer, format: int64 } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Sort' }
      responses:
        '200':
          description: A page of recipes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageRecipeResponse' }
    post:
      tags: [recipes]
      summary: Create a recipe (tags are get-or-create by name)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecipeRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }

  /api/recipes/by-ids:
    get:
      tags: [recipes]
      summary: Batch lookup by id (cross-service resolution; unknown ids omitted)
      parameters:
        - name: id
          in: query
          required: true
          description: Repeatable; at most 500.
          schema: { type: array, items: { type: integer, format: int64 } }
      responses:
        '200':
          description: The recipes that exist, in no particular order
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RecipeResponse' }
        '400': { $ref: '#/components/responses/Problem' }

  /api/recipes/random:
    get:
      tags: [recipes]
      summary: One random recipe, optionally from a filtered set (same filters as the list)
      parameters:
        - { name: name, in: query, schema: { type: string } }
        - name: tag
          in: query
          description: Repeatable. Tag names to filter by.
          schema: { type: array, items: { type: string } }
        - name: match
          in: query
          schema: { type: string, enum: [all, any], default: all }
        - name: ingredientId
          in: query
          description: Repeatable.
          schema: { type: array, items: { type: integer, format: int64 } }
      responses:
        '200':
          description: A random matching recipe
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeResponse' }
        '404': { $ref: '#/components/responses/Problem' }

  /api/recipes/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [recipes]
      summary: Get one recipe
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeResponse' }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      tags: [recipes]
      summary: Replace a recipe (steps, ingredients, tags, and optionally name/creator/version)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecipeRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [recipes]
      summary: Delete a recipe
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Problem' }

  /api/recipes/{id}/tags:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    post:
      tags: [recipes]
      summary: Attach tags (get-or-create by name) to a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagNamesRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }

  /api/recipes/{id}/tags/{tagName}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
      - name: tagName
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [recipes]
      summary: Detach one tag from a recipe (the tag itself is kept)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeResponse' }
        '404': { $ref: '#/components/responses/Problem' }

  /api/recipes/import:
    post:
      tags: [recipes]
      summary: Bulk-create recipe skeletons from a CSV upload
      description: >
        `multipart/form-data`, part `file`. Header row required; columns
        (case-insensitive): `name` (required), `creator`, `tags`
        (`;`-separated), `ingredients` (`;`-separated `<id>` /
        `<id>:<amount>:<unit>` / `<id>:<amount>:<unit>:opt` tokens). Steps are
        not imported. Always 200 — per-row problems are in the body.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: Import report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportResult' }
        '400': { $ref: '#/components/responses/Problem' }

  /api/tags:
    get:
      tags: [tags]
      summary: List / autocomplete tags
      parameters:
        - name: prefix
          in: query
          description: Name starts with this (autocomplete).
          schema: { type: string }
        - name: namespace
          in: query
          schema: { type: string }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Sort' }
      responses:
        '200':
          description: A page of tags
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageTagResponse' }
    post:
      tags: [tags]
      summary: Create a tag explicitly (with namespace / description)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }

  /api/tags/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [tags]
      summary: Get one tag
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagResponse' }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      tags: [tags]
      summary: Rename a tag / set its namespace or description
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [tags]
      summary: Delete a tag (detaches it from every recipe first)
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Problem' }

  /api/tags/merge:
    post:
      tags: [tags]
      summary: Fold one or more tags into another, moving every recipe link
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MergeRequest' }
      responses:
        '200':
          description: The surviving tag
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }

components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, default: 20 }
    Sort:
      name: sort
      in: query
      description: '`property,(asc|desc)` — e.g. `name,asc`.'
      schema: { type: string }

  responses:
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }

  schemas:
    RecipeRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 255 }
        creator: { type: string, nullable: true, description: Defaults to "anonymous" when null. }
        version:
          type: integer
          nullable: true
          description: Assigned automatically (next for this name+creator) when null.
        steps:
          type: array
          items: { $ref: '#/components/schemas/RecipeStep' }
        ingredients:
          type: array
          items: { $ref: '#/components/schemas/RecipeIngredientInput' }
        tags:
          type: array
          items: { type: string }
          description: Tag names; created if they don't exist.
    RecipeResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        creator: { type: string }
        version: { type: integer }
        steps:
          type: array
          items: { $ref: '#/components/schemas/RecipeStep' }
        ingredients:
          type: array
          items: { $ref: '#/components/schemas/RecipeIngredientOutput' }
        tags:
          type: array
          items: { $ref: '#/components/schemas/TagResponse' }
    RecipeStep:
      type: object
      required: [text]
      properties:
        position: { type: integer, description: Output only — the step's index. Ignored on input. }
        text: { type: string }
        tools:
          type: array
          items: { type: string }
          description: Implements this step needs (e.g. "wok", "candy thermometer").
    RecipeIngredientInput:
      type: object
      required: [ingredientId]
      properties:
        ingredientId: { type: integer, format: int64, description: Id in the Ingredient catalogue. }
        quantity: { type: string, nullable: true, description: 'Free text for display ("2 cups, packed", "a good pinch").' }
        amount: { type: number, nullable: true, minimum: 0, description: 'Machine-readable quantity; requires `unit`. Feeds the calculators.' }
        unit:
          type: string
          nullable: true
          description: >
            Unit for `amount` (lenient: `g`/`grams`, `kg`, `mg`, `oz`, `lb`,
            `ml`, `l`, `tsp`/`teaspoon`, `tbsp`/`tablespoon`, `cup`, `fl_oz`,
            `piece`, `clove`, `slice`, `pinch`). Must be present iff `amount` is.
        optional: { type: boolean, default: false }
        replaceable: { type: boolean, default: false, description: 'If true, replacements must be non-empty.' }
        replacements:
          type: array
          items: { $ref: '#/components/schemas/Replacement' }
    # RecipeIngredientInput + the server-assigned id. Flattened (not `allOf`) so
    # strict validators don't treat each branch as a closed object.
    RecipeIngredientOutput:
      type: object
      required: [ingredientId]
      properties:
        id: { type: integer, format: int64 }
        ingredientId: { type: integer, format: int64, description: Id in the Ingredient catalogue. }
        quantity: { type: string, nullable: true, description: 'Free text for display.' }
        amount: { type: number, nullable: true, description: 'Machine-readable quantity, paired with `unit`.' }
        unit: { type: string, nullable: true, description: 'Lower-cased unit token, e.g. `g`, `cup`, `tbsp`.' }
        optional: { type: boolean, default: false }
        replaceable: { type: boolean, default: false, description: 'If true, replacements must be non-empty.' }
        replacements:
          type: array
          items: { $ref: '#/components/schemas/Replacement' }
    Replacement:
      type: object
      required: [ingredientId]
      properties:
        ingredientId: { type: integer, format: int64, description: The substitute catalogue entry. }
        recipeId:
          type: integer
          format: int64
          nullable: true
          description: A recipe here that makes the substitute, if it can be made rather than bought.
    TagRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 100 }
        namespace: { type: string, maxLength: 60, nullable: true }
        description: { type: string, maxLength: 500, nullable: true }
    ImportResult:
      type: object
      properties:
        rows:     { type: integer, description: Data rows seen (header excluded). }
        imported: { type: integer }
        skipped:  { type: integer }
        errors:
          type: array
          items:
            type: object
            properties:
              line:    { type: integer, format: int64, description: 1-based line number in the file. }
              message: { type: string }
    TagResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, description: Normalised to trimmed lower-case. }
        namespace: { type: string, nullable: true }
        description: { type: string, nullable: true }
    TagNamesRequest:
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          minItems: 1
          items: { type: string }
    MergeRequest:
      type: object
      required: [from, into]
      properties:
        from:
          type: array
          minItems: 1
          items: { type: integer, format: int64 }
        into: { type: integer, format: int64 }
    ProblemDetail:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: object
          additionalProperties: { type: string }
          description: Present on 400 validation failures — field name to message.
    # Stable page envelope (`PageResponse<T>`) — hand-controlled, not Spring
    # Data's internal `Page` serialization. Flattened (not `allOf` composed) so
    # strict validators don't treat each branch as a closed object.
    PageRecipeResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, first, last, numberOfElements, empty]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/RecipeResponse' }
        page: { type: integer, description: Zero-based index of this page. }
        size: { type: integer, description: Requested page size. }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
        numberOfElements: { type: integer, description: Items actually on this page. }
        empty: { type: boolean }
    PageTagResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, first, last, numberOfElements, empty]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/TagResponse' }
        page: { type: integer, description: Zero-based index of this page. }
        size: { type: integer, description: Requested page size. }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
        numberOfElements: { type: integer, description: Items actually on this page. }
        empty: { type: boolean }